Refactor more adapter files and update include paths
