+++ src/control/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>long_control</name>
+  <name>control</name>
   <version>0.0.0</version>
   <description>TODO: Package description</description>
   <maintainer email="louren-o@todo.todo">louren-o</maintainer>
