+++ src/planning/CMakeLists.txt
@@ -31,9 +31,9 @@ set(TESTFILES
 )
 
 set(IMPLEMENTATION_FILES
-  src/adapter/adapter.cpp
-  src/adapter/eufs.cpp
-  src/adapter/fsds.cpp
+  src/adapter_planning/adapter.cpp
+  src/adapter_planning/eufs.cpp
+  src/adapter_planning/fsds.cpp
   src/planning/global_path_planner.cpp
   src/planning/local_path_planner.cpp
   src/planning/track.cpp
