+++ src/planning/include/adapter_planning/fsds.hpp
@@ -1,7 +1,7 @@
 #ifndef SRC_PLANNING_PLANNING_INCLUDE_ADAPTER_FSDS_HPP_
 #define SRC_PLANNING_PLANNING_INCLUDE_ADAPTER_FSDS_HPP_
 
-#include "adapter/adapter.hpp"
+#include "adapter_planning/adapter.hpp"
 
 class FsdsAdapter : public Adapter {
   rclcpp::Subscription<fs_msgs::msg::GoSignal>::SharedPtr fsds_state_subscription_;
