+++ src/planning/src/adapter_planning/eufs.cpp
@@ -1,4 +1,4 @@
-#include "adapter/eufs.hpp"
+#include "adapter_planning/eufs.hpp"
 
 #include "planning/planning.hpp"
