+++ src/planning/src/planning/planning.cpp
@@ -5,9 +5,9 @@
 #include <string>
 #include <vector>
 
-#include "adapter/eufs.hpp"
-#include "adapter/fsds.hpp"
-#include "adapter/map.hpp"
+#include "adapter_planning/eufs.hpp"
+#include "adapter_planning/fsds.hpp"
+#include "adapter_planning/map.hpp"
 #include "custom_interfaces/msg/cone_array.hpp"
 #include "custom_interfaces/msg/path_point.hpp"
 #include "custom_interfaces/msg/path_point_array.hpp"
